refactor[#137] : NavigationBar 모두 shared 로 이동
