+++ src/shared/ui/navigationbar/GlobalNavigationBar.tsx
