+++ src/shared/ui/navigationbar/index.ts
@@ -1,1 +1,3 @@
-export { default as NavigationBar } from "./NavigationBar";
+export * from "./BackwardNavigationBar";
+export * from "./CloseNavigationBar";
+export * from "./GlobalNavigationBar";
