--- src/widgets/navigationbar/ui/index.ts
@@ -1,3 +1,0 @@
-export { default as BackwardNavigationBar } from "./BackwardNavigationBar";
-export * from "./CloseNavigationBar";
-export * from "./GlobalNavigationBar";
